refactor: reorganize configuration files and enhance database setup
