+++ src/config/server/websocket.ts
@@ -1,9 +1,9 @@
 import http from 'http';
 import WebSocket from 'ws';
-import { MessageDb } from '../database/messageDb';
-import { MessageService } from '../services/messageService';
-import { WebSocketService } from '../services/websocketService';
-import Logger from '../utils/logger';
+import { MessageDb } from '../../database/messageDb';
+import { MessageService } from '../../services/messageService';
+import { WebSocketService } from '../../services/websocketService';
+import { withErrorHandler } from '../../utils/asyncHandler';
 
 /**
  * Sets up a WebSocket server using the provided HTTP server.
@@ -12,8 +12,8 @@ import Logger from '../utils/logger';
  * @returns The configured WebSocket server instance.
  * @throws Will throw an error if the WebSocket server setup fails.
  */
-export function setupWebSocket(server: http.Server): WebSocket.Server {
-  try {
+export function setupWebSocket(server: http.Server): Promise<WebSocket.Server> {
+  return withErrorHandler(async () => {
     const wss = new WebSocket.Server({ server });
     const messageDb = new MessageDb();
     const messageService = new MessageService(messageDb);
@@ -22,10 +22,6 @@ export function setupWebSocket(server: http.Server): WebSocket.Server {
     wss.on('connection', (ws: WebSocket) => {
       websocketService.handleConnection(ws);
     });
-
     return wss;
-  } catch (error) {
-    Logger.error('Failed to setup WebSocket server:', error as Error);
-    throw error;
-  }
+  }, 'Failed to setup WebSocket server:');
 }
